fix some overflow bugs which caused various sizes to be misreported

+++ buyvmmanager/BVMHumanValueTransformer.h
@@ -2,6 +2,6 @@
 
 @interface BVMHumanValueTransformer : NSObject
 
-+ (NSString *)humanSizeValueFromBytes:(NSUInteger)bytes;
++ (NSString *)humanSizeValueFromBytes:(long long)bytes;
 
 @end
